refactor(`server/indexer/dir_entry_guesser/tests`): rm unnecessary cast

+++ server/main/indexer/dir_entry_guesser/tests.rs
@@ -24,7 +24,7 @@ mod tests {
                 last_modified: Some(DateTime::<Utc>::from(
                     $file.metadata().unwrap().modified().unwrap(),
                 )),
-                size: Some($file.metadata().unwrap().len() as i64),
+                size: Some($file.metadata().unwrap().len()),
             }
         };
     }
